17-04-2024

+++ xenia_and_ringroad.cpp
@@ -2,14 +2,14 @@
 using namespace std;
 vector<int>houses;
 int main(){
-    int a,b;
+    long long a,b;
     cin >> a >> b;
     for(int i = 0 ; i<b ; i++){
         int x;
         cin >> x;
         houses.push_back(x);
     }
-    int ans=houses[0]-1;
+    long long ans=houses[0]-1;
     for(int i=1 ; i<b ; i++){
         if(houses[i]>=houses[i-1]){
             ans+=houses[i]-houses[i-1];
